Add straightforward casts for %o and %x
clang19 -Wformat

- archive_entry_mode() returns a __LA_MODE_T, which is either an int,
  unsigned short, or mode_t (which is an "integer type", but POSIX
  doesn't specify signed-ness or length).

  The casts are obviously safe for (int) or (unsigned short).  In the
  vanishingly unlikely case of mode_t being a longer integer, these
  particular casts are still ok because the value is always &0777, so
  we wouldn't be truncating any info.

- myacls[].type and .permset are both ints (struct archive_tst_acl_l is
  defined in tests_utils/test_common.h), so casting to (unsigned int) is
  safe.

- the 'type' and 'permset' in libarchive/test/test_acl_platform_nfs4.c
  are also (int), so again it's safe.

+++ test_utils/test_main.c
@@ -3319,7 +3319,8 @@ assertion_entry_set_acls(const char *file, int line, struct archive_entry *ae,
 			ret = 1;
 			failure_start(file, line, "type=%#010x, "
 			    "permset=%#010x, tag=%d, qual=%d name=%s",
-			    acls[i].type, acls[i].permset, acls[i].tag,
+			    (unsigned int)acls[i].type,
+			    (unsigned int)acls[i].permset, acls[i].tag,
 			    acls[i].qual, acls[i].name);
 			failure_finish(NULL);
 		}
@@ -3451,7 +3452,8 @@ assertion_entry_compare_acls(const char *file, int line,
 		} else if (matched != 1) {
 			failure_start(file, line, "Could not find match for "
 			    "ACL (type=%#010x,permset=%#010x,tag=%d,qual=%d,"
-			    "name=``%s'')", type, permset, tag, qual, name);
+			    "name=``%s'')", (unsigned int)type,
+			    (unsigned int)permset, tag, qual, name);
 			failure_finish(NULL);
 			ret = 1;
 		}
@@ -3471,7 +3473,8 @@ assertion_entry_compare_acls(const char *file, int line,
 	if (n != 0) {
 		failure_start(file, line, "Could not find match for ACL "
 		    "(type=%#010x,permset=%#010x,tag=%d,qual=%d,name=``%s'')",
-		    acls[marker[0]].type, acls[marker[0]].permset,
+		    (unsigned int)acls[marker[0]].type,
+		    (unsigned int)acls[marker[0]].permset,
 		    acls[marker[0]].tag, acls[marker[0]].qual,
 		    acls[marker[0]].name);
 		failure_finish(NULL);
